+++ backend/src/routes/images.router.js
@@ -2,7 +2,7 @@ import express from "express";
 import { deleteImage, getImage, getImages, uploadImage } from "../controllers/images.controller.js";
 import { paramIdValidator } from "../utils/validator.manager.js";
 import { uploadMiddleware } from "../middlewares/upload.middleware.js";
-import { requireJwt } from "../middlewares/require.token.js";
+import { requireJwt } from "../middlewares/token.middleware.js";
 
 const router = express.Router();
